pyproject.toml
[tool.poetry]
name = "articles"
version = "0.1.0"
description = ""
authors = ["shu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-firestore = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
